removed BoardPrinter.java, moved method inside new class ConsoleOutputHandler
